Changed the visitor design to increase the flexibility of the kind of visitation rule that can be used.
Before, the rules looked like this
    Env1Args2Blocks $ \(a_string, a_raw_block, another_raw_block) state -> do ...

where Env1Args2Blocks is one case of an algebraic type. This means that there has to be one case for each for each possible list of arguments of LaTeX environmants and commands with a corresponding branch in the visitor algorithm. Now, all the possible argument lists are covered by one case for environments and for commands.

All the magic resides in Document.TypeList. It defines two classes: Readable and TypeList. Every type that can be parsed as the argument of an environment or a command must be an instance of Readable. It provides an action that convert a LaTeX parse tree into one argument of the proper type. TypeList provides a command that parses a tuple of Readable arguments.

Everything is statically typed.