Changed the schedule system so that every event has a set of schedule and a sequence of schedule refinement rules.
o \transient now take a schedule number as a second argument.
o it is now possible (and necessary) to use \weakento{evt}{3}{remove}{add} to introduce schedules to event that didn't use to have one
o it is now possible (but untested) to use \replace{evt}{3}{remove}{add}{keep}{prog}{saf} in order to replace parts of a schedule and prove the soundness using prog and saf, the leads-to property and the safety property that needs to be proved.