Decomposed feasibility proof obligations into independant clauses. The former formulation for an assignment to two variables was: (∃ x',y':: x' = E ∧ y' = F) but now this will be two separate proof obligations: (∃ x':: x' = E) and (∃ y':: y' = F) unless an assignment refers to more than one primed variable.
It proceeds by partitioning primed variables and expressions into as many partition as possible so that expressions are in the same partition as all the primed variables it refers to. Each partition then form an existential quantification.

Finally, any assignment of the form P ∧ Q is broken down into two assignments.