Display the proof result in a colour-changing bar
o Refactor ProofResult and manually write a ToJSON instance to add
css classes for alerts and glyphicons depending on the result of
verification.

o Use the <noscript> tag for displaying the warning to users that have
disabled JavaScript, instead of manually deleting the tag through JS.
This way, we let the browser take care of displaying it only when
JS is disabled. It's much faster this way and there's no flicker.

o The new "result bar" will fade in once the julius file is loaded,
and when clicking prove, it shows the result of verification along with
a suiting colour and icon.