Add interval_theory when arithmetic is present
o Excludes interval_theory when making a list of supported theories for
sending to the front-end.

o When receiving a proof form, checks and if arithmetic was selected by
the user, it adds interval_theory as well.

o Use the invariants package and add Pre to the context of getTheory,
so that if the precondition fails (the given string is not a valid
theory name), we'll get an error message pointing to the source of error