Merge Presets and Examples and categorize them.
o Merge presets and examples together. The formerly "Presets" button is
now "Examples ↓" which will smoothly scroll down to the Examples panel.

o Examples are now organized into categories (theories), with an extra
category in the end where we can provide custom examples (outside of
presets).

TODO:

o LaTeX rendering for preset goals.

o [OPTIONAL]: Figure out a way to animate scroll to top when clicking a
"Try" button, without flickering. Right now, since we "click" the + or -
buttons, they each focus one of the text inputs after they're done. Make
this parametric so that it only happens when the user clicks them and
not when we simulate their click so we can smoothly animate to top,
without focus events causing flickers.