[WIP][Subtyping] Add t component to FunApp
o Begin working on subtyping: add stub Logic.Expr.Subtyping module
(currently not compiling, and not in the build system).

o Add a [failing] test case (case25) for subtyping.

o Modify the Show instance of ExprNotation to display both the Context
and the ExprNotation (Notation & Expr).

o Bring in preludeTheories when using theory_setting' for inclusion of a
table of theories.

o Bump stack snapshot to lts-6.14.

o Remove the <4.14 upper bound for lens, and explicitly use lens-4.14 in
stack.